+++ src/containers/user-profile/comments-with-rating-block/rating-block/RatingBlock.styles.ts
