+++ src/pages/user-profile/UserProfile.jsx
@@ -2,10 +2,10 @@ import { useCallback } from 'react'
 import { useAppSelector } from '~/hooks/use-redux'
 import { useParams, useSearchParams, useMatch } from 'react-router-dom'
 
-import ProfileInfo from '~/containers/tutor-profile/profile-info/ProfileInfo'
+import ProfileInfo from '~/containers/user-profile/profile-info/ProfileInfo'
 import CompleteProfileBlock from '~/components/complete-profile/CompleteProfileBlock'
-import VideoPresentation from '~/containers/tutor-profile/video-presentation/VideoPresentation'
-import CommentsWithRatingBlock from '~/containers/tutor-profile/comments-with-rating-block/CommentsWithRaitngBlock'
+import VideoPresentation from '~/containers/user-profile/video-presentation/VideoPresentation'
+import CommentsWithRatingBlock from '~/containers/user-profile/comments-with-rating-block/CommentsWithRatingBlock'
 import PageWrapper from '~/components/page-wrapper/PageWrapper'
 import Loader from '~/components/loader/Loader'
 import { userService } from '~/services/user-service'
@@ -14,13 +14,13 @@ import videoImgProfile from '~/assets/img/tutor-profile-page/presentationVideoIm
 
 import { profileItems } from '~/components/profile-item/complete-profile.constants'
 import { defaultResponses } from '~/constants'
-import { responseMock } from '~/pages/tutor-profile/constants'
-import AboutTutorBlock from '~/containers/tutor-profile/about-tutor-block/AboutTutorBlock'
+import { responseMock } from '~/pages/user-profile/constants'
+import AboutTutorBlock from '~/containers/user-profile/about-tutor-block/AboutTutorBlock'
 
 import { UserRoleEnum } from '~/types'
 import { authRoutes } from '~/router/constants/authRoutes'
 
-const TutorProfile = () => {
+const UserProfile = () => {
   const { id } = useParams()
   const [searchParams] = useSearchParams()
   const { userId, userRole } = useAppSelector((state) => state.appMain)
@@ -77,4 +77,4 @@ const TutorProfile = () => {
   )
 }
 
-export default TutorProfile
+export default UserProfile
